+++ merkle/compact/range_test.go
@@ -26,7 +26,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	rfc6962 "github.com/google/trillian/merkle/rfc6962/hasher"
+	"github.com/google/trillian/merkle/rfc6962"
 	"github.com/google/trillian/merkle/testonly"
 
 	_ "github.com/golang/glog" // Required for flag handling